split __main__.py from __init__.py
